+++ app/Http/Controllers/ServiceController.php
@@ -26,6 +26,8 @@ class ServiceController extends Controller
      */
     public function create()
     {
+        $this->authorize("service-create", Service::class);
+
         return view("backoffice.service.create");
 
     }
@@ -38,6 +40,8 @@ class ServiceController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize("create", Service::class);
+
         
         $request->validate([
             "logo"=>"required",
@@ -75,6 +79,8 @@ class ServiceController extends Controller
      */
     public function edit(Service $service)
     {
+        $this->authorize("service-edit", Service::class);
+
        return view("backoffice.service.edit", compact("service"));
 
     }
@@ -88,6 +94,8 @@ class ServiceController extends Controller
      */
     public function update(Request $request, Service $service)
     {
+        $this->authorize("update", $service);
+
         $request->validate([
             "logo"=>"required",
             "petitTitre"=>"required",
@@ -111,6 +119,8 @@ class ServiceController extends Controller
      */
     public function destroy(Service $service)
     {
+        $this->authorize("delete", $service);
+
         $service->delete();
         return redirect()->back();
     }
